Server/requirements.txt
mysql-connector-python
email-validator
requests